🚚 Move Munkres to Core module
