--- modules/assignment/munkres/CMakeLists.txt
@@ -1,49 +1,0 @@
-#############################################################################
-#
-# ViSP, open source Visual Servoing Platform software.
-# Copyright (C) 2005 - 2019 by Inria. All rights reserved.
-#
-# This software is free software; you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation; either version 2 of the License, or
-# (at your option) any later version.
-# See the file LICENSE.txt at the root directory of this source
-# distribution for additional information about the GNU GPL.
-#
-# For using ViSP with software that can not be combined with the GNU
-# GPL, please contact Inria about acquiring a ViSP Professional
-# Edition License.
-#
-# See http://visp.inria.fr for more information.
-#
-# This software was developed at:
-# Inria Rennes - Bretagne Atlantique
-# Campus Universitaire de Beaulieu
-# 35042 Rennes Cedex
-# France
-#
-# If you have questions regarding the use of this file, please contact
-# Inria at visp@inria.fr
-#
-# This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
-# WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
-#
-# Description:
-# ViSP assignment algorithm files.
-#
-# Authors:
-# Fabien Spindler
-#
-#############################################################################
-
-if(VISP_CXX_STANDARD GREATER_EQUAL VISP_CXX_STANDARD_17)
-    vp_add_module(munkres visp_core)
-    vp_glob_module_sources()
-    vp_module_include_directories()
-    vp_create_module()
-
-    if(WITH_CATCH2)
-        include_directories(${CATCH2_INCLUDE_DIRS})
-    endif()
-    vp_add_tests()
-endif()
